slight cleanup of client tests
